requirements.txt
accelerate>=1.13.0
datasets>=4.8.4
numpy>=2.4.4
pandas>=3.0.2
PyYAML>=6.0.3
torch>=2.11.0
tqdm>=4.67.3
transformers>=5.5.4
wandb>=0.26.0
python-dotenv>=1.2.2
kagglehub>=1.0.0
